Move SetPZ69DisplayBytes to PZ69DisplayBytes class
